+++ trace.h
@@ -1,7 +1,7 @@
 #if !defined(__POLYRAY_TRACE_DEFS)
 #define __POLYRAY_TRACE_DEFS
 
-extern float Trace(Viewpoint *, int, Flt, Ray *, Vec, Flt *, Flt, unsigned long *);
+extern float Trace(Viewpoint *, int, Flt, Ray *, Vec, Flt *, Flt, unsigned int *);
 
 #endif /* __POLYRAY_TRACE_DEFS */
